Se implementaron mas estados y un visitor para asignar estados
En particular, el visitor se encargaria de asignar estados a los
aliados. Todavia no se han creado objetos que generen un estado.

+++ TowerDefenseTDP2019/src/state/Estado.java
@@ -2,10 +2,10 @@ package state;
 
 public abstract class Estado {
 	
-	public abstract int getDaño(int dañoBase);
+	public abstract float getDaño(float dañoBase);
 	
 	public abstract int getVelocidad(int velBase);
 	
-	public abstract int recibirDaño(int dañoBase);
+	public abstract float recibirDaño(float dañoBase);
 
 }
